+++ src/translations/de/facebook-connector.php
@@ -22,4 +22,6 @@
  */
 return [
     'FacebookConnector plugin loaded' => 'FacebookConnector plugin loaded',
+    'Settings' => 'Einstellungs',
+    'info' => 'Infos'
 ];