+++ CHANGELOG.md
@@ -4,12 +4,32 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## 0.7.0 - 2017-09-22
+## v0.11.0 - 2017-09-25
+### Changed
+  - Moved the token callback function to the dashboard page, so no more templates are needed
+
+## v0.11.0 - 2017-09-25
 ### Added
-  - Implemented OAtuh Login
+  - Boilerplate code added for the widget
+
+## v0.10.0 - 2017-09-25
+### Added
+  - Entry field options can no be overwritten in the file fieldconfig.php
+  
+## v0.9.0 - 2017-09-24
+### Added
+  - Options can now be defined in the plugin settings
+
+## v0.8.0 - 2017-09-24
+### Changed
+   - Changed the readme and added som information's about the project
+
+## v0.7.0 - 2017-09-22
+### Added
+  - Implemented OAuth Login
   - Implemented post to facebook
   - Implemented update a post on facebook
 
-## 0.1.0 - 2017-09-16 (Initial Release)
+## v0.1.0 - 2017-09-16 (Initial Release)
 ### Added
  - Initial release